add spec for readme examples
The readme examples now have their own spec to ensure their
correctness. The spec is included in the SpecRunner.
The name property had to be renamed. For more information see #1.

The spec exposed a bug, which is fixed in this commit. Placeholders
were only replaced once instead of every time the function was
called with parameters. The function now pushes to a new array, which
is used to call the mapped method.